fix(chaos): add tests, validation and trajectory improvements for ChebyshevLogistic2d
This commit addresses review feedback for PR #459:
- Reverted unintentional changes to `export.json` and added it to `.gitignore`.
- Added `debug_assert!` validation for `k` and `a` parameters to prevent degenerate states.
- Improved the hyperchaotic trajectory logic by introducing `sin(pi * ...)` coupling and periodic perturbations, ensuring sensitivity to initial conditions and preventing state collapse (as verified by new tests).
- Added comprehensive unit tests for range, determinism, sensitivity, non-degeneracy, and statistical distribution.
- Properly exported `ChebyshevLogistic2d` in `csm-chaos` and re-exported it in `csm-core`.
- Added clarifying comments to the `chaotic_hashing_chebyshev` benchmark.

What: Robust implementation of 2D Chebyshev-Logistic Hyperchaotic Map.
Why: Ensure mathematical correctness, prevent state collapse, and provide full test coverage.
Impact: High-entropy chaotic source for LSH with verified sensitivity and non-degeneracy.
