+++ .github/workflows/pre-release-gate.yml
@@ -117,7 +117,7 @@ jobs:
         cache: false
 
     - name: Install cargo-mutants
-      uses: taiki-e/install-action@16b05812d776ae1dfaabc8277e421fb6d2506419  # v2.82.7
+      uses: taiki-e/install-action@9e1e5806d4a4822de933115878265be9aaa786d9  # v2.82.2
       with:
         tool: cargo-mutants
