+++ VERSION
@@ -1,1 +1,1 @@
-0.3.7
+0.3.6
