+++ crates/csm-core/src/maps/mod.rs
@@ -1,1 +1,2 @@
 pub mod hyperchaotic;
+pub mod hyperchaotic_chebyshev;
