+++ crates/csm-memory/src/singularity.rs
@@ -491,9 +491,7 @@ impl<H: Hypervector + 'static> Singularity<H> {
                 }
             }
         }
-        // Optimization: Use sort_unstable_by to avoid stability overhead for numeric ranking.
-        incoming
-            .sort_unstable_by(|a, b| b.1.partial_cmp(&a.1).unwrap_or(std::cmp::Ordering::Equal));
+        incoming.sort_by(|a, b| b.1.partial_cmp(&a.1).unwrap_or(std::cmp::Ordering::Equal));
         incoming
     }
