+++ crates/csm-retrieval/Cargo.toml
@@ -8,7 +8,6 @@ license = "MIT"
 repository = "https://github.com/d-o-hub/chaotic_semantic_memory"
 
 [dependencies]
-chrono = { workspace = true }
 csm-core = { path = "../csm-core", version = "0.3.7" }
 csm-memory = { path = "../csm-memory", version = "0.3.7" }
 serde = { workspace = true }
