+++ crates/csm-retrieval/src/lib.rs
@@ -13,10 +13,7 @@ mod rerank;
 
 pub use bm25::{Bm25Config, Bm25Index};
 pub use graph_rag::{GraphRagConfig, GraphRagResult, graph_rag_retrieve};
-pub use hybrid::{
-    HybridConfig, HybridMode, HybridResult, RetrievalAbstention, merge_results,
-    merge_results_checked, normalize_scores,
-};
+pub use hybrid::{HybridConfig, HybridMode, merge_results, normalize_scores};
 pub use rerank::{MmrReranker, RecencyDecayReranker, RerankCandidate, Reranker, parse_rerankers};
 
 #[cfg(feature = "rerank-cross")]
