+++ llms.txt
@@ -12,7 +12,6 @@
 - async-trait (0.1.83)
 - base64
 - bincode (1.3.3)
-- chrono
 - clap (4.5) [features: derive, env, string]
 - clap_complete (4.5.42)
 - cloudevents-sdk (0.9.0)
@@ -498,6 +497,7 @@ Created by: [cargo-llms-txt](https://github.com/masinc/cargo-llms-txt)
 - impl McpHandler
 - impl McpHandler
 - impl ServerHandler for McpHandler
+- impl McpHandler
 
 ### src/mcp/mod.rs
 
@@ -527,10 +527,6 @@ Created by: [cargo-llms-txt](https://github.com/masinc/cargo-llms-txt)
 - impl Default for McpConfig
 - pub fn serve
 - impl Clone for McpHandler
-
-### src/mcp/tools.rs
-
-- impl McpHandler
 
 ### src/metadata_filter.rs
 
@@ -626,12 +622,9 @@ Created by: [cargo-llms-txt](https://github.com/masinc/cargo-llms-txt)
 
 ### src/retrieval/hybrid.rs
 
-- pub struct RetrievalAbstention
-- pub enum HybridResult
 - pub fn compute_weights
 - pub fn normalize_scores
 - pub fn merge_results
-- pub fn merge_results_checked
 - pub enum HybridMode
 - pub struct HybridConfig
 - impl Default for HybridConfig
@@ -644,7 +637,7 @@ Created by: [cargo-llms-txt](https://github.com/masinc/cargo-llms-txt)
 - pub mod rerank
 - pub use bm25::{Bm25Config, Bm25Index}
 - pub use graph_rag::{GraphRagConfig, GraphRagResult, graph_rag_retrieve}
-- pub use hybrid::{HybridConfig, HybridMode, HybridResult, RetrievalAbstention, compute_weights, merge_results, merge_results_checked, normalize_scores}
+- pub use hybrid::{HybridConfig, HybridMode, compute_weights, merge_results, normalize_scores}
 - pub use rerank::{MmrReranker, RerankCandidate, Reranker}
 
 ### src/retrieval/rerank.rs
@@ -1252,7 +1245,6 @@ repository = "https://github.com/d-o-hub/chaotic_semantic_memory"
 homepage = "https://github.com/d-o-hub/chaotic_semantic_memory"
 
 [workspace.dependencies]
-chrono = { version = "0.4.45", features = ["serde"] }
 serde = { version = "1.0.219", features = ["derive"] }
 serde_json = "1.0.150"
 thiserror = "2.0.11"
@@ -1291,7 +1283,6 @@ include = [
 ]
 
 [dependencies]
-chrono = { workspace = true }
 csm-core = { path = "crates/csm-core", version = "0.3.7" }
 csm-embedding = { path = "crates/csm-embedding", version = "0.3.7" }
 csm-memory = { path = "crates/csm-memory", version = "0.3.7" }
