+++ tests/arch_fitness.rs
@@ -6,7 +6,7 @@ use std::process::Command;
 const MAX_LOC: usize = 500;
 
 /// Known LOC exceptions in workspace crates (pre-existing, tracked for reduction).
-const LOC_EXCEPTIONS: &[&str] = &[
+const LOC_EXCEPTIONS_CRATES: &[&str] = &[
     "crates/csm-core/src/hyperdim.rs",
     "crates/csm-memory/src/singularity.rs",
     "crates/csm-memory/src/graph_traversal.rs",
@@ -58,13 +58,13 @@ fn loc_gate_src_files_under_500_lines() {
     let mut violations = Vec::new();
     for file in find_rs_files("src") {
         let loc = line_count(&file);
-        if loc > MAX_LOC && !LOC_EXCEPTIONS.iter().any(|e| file.ends_with(e)) {
+        if loc > MAX_LOC {
             violations.push(format!("{file}: {loc} lines"));
         }
     }
     assert!(
         violations.is_empty(),
-        "Files exceeding {MAX_LOC} LOC in src/ (excluding known exceptions):\n{}",
+        "Files exceeding {MAX_LOC} LOC in src/:\n{}",
         violations.join("\n")
     );
 }
@@ -84,7 +84,7 @@ fn loc_gate_workspace_crate_src_files_under_500_lines() {
     let mut violations = Vec::new();
     for file in files {
         let loc = line_count(&file);
-        if loc > MAX_LOC && !LOC_EXCEPTIONS.iter().any(|e| file.ends_with(e)) {
+        if loc > MAX_LOC && !LOC_EXCEPTIONS_CRATES.iter().any(|e| file.ends_with(e)) {
             violations.push(format!("{file}: {loc} lines"));
         }
     }
